Release 24/03/23 10:03:30

+++ src/main/java/org/sopt/makers/internal/external/MakersCrewProdClient.java
@@ -12,6 +12,6 @@ public interface MakersCrewProdClient {
     MemberCrewResponse getUserAllCrew(
             @RequestParam("page") Integer page,
             @RequestParam("take") Integer take,
-            @RequestParam("orgUserId") Integer orgUserId
+            @RequestParam("orgUserId") Long orgUserId
     );
 }
